--- Doc/Gato.html
@@ -1,167 +1,0 @@
-<!DOCTYPE HTML>
-<!-- NewPage -->
-<html lang="es">
-<head>
-<!-- Generated by javadoc (13.0.1) on Fri Jun 12 11:11:09 CEST 2020 -->
-<title>Gato</title>
-<meta http-equiv="Content-Type" content="text/html; charset=utf-8">
-<meta name="dc.created" content="2020-06-12">
-<meta name="description" content="declaration: package: examenEntornos, class: Gato">
-<meta name="generator" content="javadoc/ClassWriterImpl">
-<link rel="stylesheet" type="text/css" href="../stylesheet.css" title="Style">
-<link rel="stylesheet" type="text/css" href="../script-dir/jquery-ui.css" title="Style">
-<script type="text/javascript" src="../script.js"></script>
-<script type="text/javascript" src="../script-dir/jszip/dist/jszip.min.js"></script>
-<script type="text/javascript" src="../script-dir/jszip-utils/dist/jszip-utils.min.js"></script>
-<!--[if IE]>
-<script type="text/javascript" src="../script-dir/jszip-utils/dist/jszip-utils-ie.min.js"></script>
-<![endif]-->
-<script type="text/javascript" src="../script-dir/jquery-3.4.1.js"></script>
-<script type="text/javascript" src="../script-dir/jquery-ui.js"></script>
-</head>
-<body class="class-declaration">
-<script type="text/javascript">var pathtoroot = "../";
-loadScripts(document, 'script');</script>
-<noscript>
-<div>JavaScript is disabled on your browser.</div>
-</noscript>
-<header role="banner">
-<nav role="navigation">
-<div class="fixedNav">
-<!-- ========= START OF TOP NAVBAR ======= -->
-<div class="topNav"><a id="navbar.top">
-<!--   -->
-</a>
-<div class="skipNav"><a href="#skip.navbar.top" title="Skip navigation links">Skip navigation links</a></div>
-<a id="navbar.top.firstrow">
-<!--   -->
-</a>
-<ul class="navList" title="Navigation">
-<li><a href="package-summary.html">Package</a></li>
-<li class="navBarCell1Rev">Class</li>
-<li><a href="class-use/Gato.html">Use</a></li>
-<li><a href="package-tree.html">Tree</a></li>
-<li><a href="../deprecated-list.html">Deprecated</a></li>
-<li><a href="../index-files/index-1.html">Index</a></li>
-<li><a href="../help-doc.html">Help</a></li>
-</ul>
-</div>
-<div class="subNav">
-<div>
-<ul class="subNavList">
-<li>Summary:&nbsp;</li>
-<li>Nested&nbsp;|&nbsp;</li>
-<li>Field&nbsp;|&nbsp;</li>
-<li>Constr&nbsp;|&nbsp;</li>
-<li><a href="#method.summary">Method</a></li>
-</ul>
-<ul class="subNavList">
-<li>Detail:&nbsp;</li>
-<li>Field&nbsp;|&nbsp;</li>
-<li>Constr&nbsp;|&nbsp;</li>
-<li>Method</li>
-</ul>
-</div>
-<div class="navListSearch"><label for="search">SEARCH:</label>
-<input type="text" id="search" value="search" disabled="disabled">
-<input type="reset" id="reset" value="reset" disabled="disabled">
-</div>
-</div>
-<a id="skip.navbar.top">
-<!--   -->
-</a>
-<!-- ========= END OF TOP NAVBAR ========= -->
-</div>
-<div class="navPadding">&nbsp;</div>
-<script type="text/javascript"><!--
-$('.navPadding').css('padding-top', $('.fixedNav').css("height"));
-//-->
-</script>
-</nav>
-</header>
-<!-- ======== START OF CLASS DATA ======== -->
-<main role="main">
-<div class="header">
-<div class="subTitle"><span class="packageLabelInType">Package</span>&nbsp;<a href="package-summary.html">examenEntornos</a></div>
-<h1 title="Class Gato" class="title">Class Gato</h1>
-</div>
-<div class="contentContainer">
-<div class="inheritance" title="Inheritance Tree">java.lang.Object
-<div class="inheritance">examenEntornos.Gato</div>
-</div>
-<section class="description">
-<hr>
-<pre>public class <span class="typeNameLabel">Gato</span>
-extends java.lang.Object</pre>
-<dl>
-<dt><span class="simpleTagLabel">Author:</span></dt>
-<dd>Jaime Rabasco Ronda
- 
- La clase Gato es una clase que hace que representa el maullido de in gato</dd>
-</dl>
-</section>
-<section class="summary">
-<ul class="blockList">
-<!-- ========== METHOD SUMMARY =========== -->
-<li class="blockList">
-<section class="methodSummary"><a id="method.summary">
-<!--   -->
-</a>
-<h2>Method Summary</h2>
-<div class="inheritedList">
-<h3>Methods inherited from class&nbsp;java.lang.Object</h3>
-<a id="methods.inherited.from.class.java.lang.Object">
-<!--   -->
-</a><code>equals, getClass, hashCode, notify, notifyAll, toString, wait, wait, wait</code></div>
-</section>
-</li>
-</ul>
-</section>
-</div>
-</main>
-<!-- ========= END OF CLASS DATA ========= -->
-<footer role="contentinfo">
-<nav role="navigation">
-<!-- ======= START OF BOTTOM NAVBAR ====== -->
-<div class="bottomNav"><a id="navbar.bottom">
-<!--   -->
-</a>
-<div class="skipNav"><a href="#skip.navbar.bottom" title="Skip navigation links">Skip navigation links</a></div>
-<a id="navbar.bottom.firstrow">
-<!--   -->
-</a>
-<ul class="navList" title="Navigation">
-<li><a href="package-summary.html">Package</a></li>
-<li class="navBarCell1Rev">Class</li>
-<li><a href="class-use/Gato.html">Use</a></li>
-<li><a href="package-tree.html">Tree</a></li>
-<li><a href="../deprecated-list.html">Deprecated</a></li>
-<li><a href="../index-files/index-1.html">Index</a></li>
-<li><a href="../help-doc.html">Help</a></li>
-</ul>
-</div>
-<div class="subNav">
-<div>
-<ul class="subNavList">
-<li>Summary:&nbsp;</li>
-<li>Nested&nbsp;|&nbsp;</li>
-<li>Field&nbsp;|&nbsp;</li>
-<li>Constr&nbsp;|&nbsp;</li>
-<li><a href="#method.summary">Method</a></li>
-</ul>
-<ul class="subNavList">
-<li>Detail:&nbsp;</li>
-<li>Field&nbsp;|&nbsp;</li>
-<li>Constr&nbsp;|&nbsp;</li>
-<li>Method</li>
-</ul>
-</div>
-</div>
-<a id="skip.navbar.bottom">
-<!--   -->
-</a>
-<!-- ======== END OF BOTTOM NAVBAR ======= -->
-</nav>
-</footer>
-</body>
-</html>
