--- Doc/class-use/Gato.html
@@ -1,100 +1,0 @@
-<!DOCTYPE HTML>
-<!-- NewPage -->
-<html lang="es">
-<head>
-<!-- Generated by javadoc (13.0.1) on Fri Jun 12 11:11:09 CEST 2020 -->
-<title>Uses of Class examenEntornos.Gato</title>
-<meta http-equiv="Content-Type" content="text/html; charset=utf-8">
-<meta name="dc.created" content="2020-06-12">
-<meta name="description" content="use: package: examenEntornos, class: Gato">
-<meta name="generator" content="javadoc/ClassUseWriter">
-<link rel="stylesheet" type="text/css" href="../../stylesheet.css" title="Style">
-<link rel="stylesheet" type="text/css" href="../../script-dir/jquery-ui.css" title="Style">
-<script type="text/javascript" src="../../script.js"></script>
-<script type="text/javascript" src="../../script-dir/jszip/dist/jszip.min.js"></script>
-<script type="text/javascript" src="../../script-dir/jszip-utils/dist/jszip-utils.min.js"></script>
-<!--[if IE]>
-<script type="text/javascript" src="../../script-dir/jszip-utils/dist/jszip-utils-ie.min.js"></script>
-<![endif]-->
-<script type="text/javascript" src="../../script-dir/jquery-3.4.1.js"></script>
-<script type="text/javascript" src="../../script-dir/jquery-ui.js"></script>
-</head>
-<body class="class-use">
-<script type="text/javascript">var pathtoroot = "../../";
-loadScripts(document, 'script');</script>
-<noscript>
-<div>JavaScript is disabled on your browser.</div>
-</noscript>
-<header role="banner">
-<nav role="navigation">
-<div class="fixedNav">
-<!-- ========= START OF TOP NAVBAR ======= -->
-<div class="topNav"><a id="navbar.top">
-<!--   -->
-</a>
-<div class="skipNav"><a href="#skip.navbar.top" title="Skip navigation links">Skip navigation links</a></div>
-<a id="navbar.top.firstrow">
-<!--   -->
-</a>
-<ul class="navList" title="Navigation">
-<li><a href="../package-summary.html">Package</a></li>
-<li><a href="../Gato.html" title="class in examenEntornos">Class</a></li>
-<li class="navBarCell1Rev">Use</li>
-<li><a href="../package-tree.html">Tree</a></li>
-<li><a href="../../deprecated-list.html">Deprecated</a></li>
-<li><a href="../../index-files/index-1.html">Index</a></li>
-<li><a href="../../help-doc.html">Help</a></li>
-</ul>
-</div>
-<div class="subNav">
-<div class="navListSearch"><label for="search">SEARCH:</label>
-<input type="text" id="search" value="search" disabled="disabled">
-<input type="reset" id="reset" value="reset" disabled="disabled">
-</div>
-</div>
-<a id="skip.navbar.top">
-<!--   -->
-</a>
-<!-- ========= END OF TOP NAVBAR ========= -->
-</div>
-<div class="navPadding">&nbsp;</div>
-<script type="text/javascript"><!--
-$('.navPadding').css('padding-top', $('.fixedNav').css("height"));
-//-->
-</script>
-</nav>
-</header>
-<main role="main">
-<div class="header">
-<h1 title="Uses of Class examenEntornos.Gato" class="title">Uses of Class<br>examenEntornos.Gato</h1>
-</div>
-<div class="classUseContainer">No usage of examenEntornos.Gato</div>
-</main>
-<footer role="contentinfo">
-<nav role="navigation">
-<!-- ======= START OF BOTTOM NAVBAR ====== -->
-<div class="bottomNav"><a id="navbar.bottom">
-<!--   -->
-</a>
-<div class="skipNav"><a href="#skip.navbar.bottom" title="Skip navigation links">Skip navigation links</a></div>
-<a id="navbar.bottom.firstrow">
-<!--   -->
-</a>
-<ul class="navList" title="Navigation">
-<li><a href="../package-summary.html">Package</a></li>
-<li><a href="../Gato.html" title="class in examenEntornos">Class</a></li>
-<li class="navBarCell1Rev">Use</li>
-<li><a href="../package-tree.html">Tree</a></li>
-<li><a href="../../deprecated-list.html">Deprecated</a></li>
-<li><a href="../../index-files/index-1.html">Index</a></li>
-<li><a href="../../help-doc.html">Help</a></li>
-</ul>
-</div>
-<a id="skip.navbar.bottom">
-<!--   -->
-</a>
-<!-- ======== END OF BOTTOM NAVBAR ======= -->
-</nav>
-</footer>
-</body>
-</html>
